Merged PR 18355: Switched from implicit to code flow with pkce
Switched from implicit to code flow with pkce

# Testing

This is not part of build and cannot be deployed as-is. The simplest way it to build and run locally on developer machine.

Update security element in app settings to point to an EG shell installation and a valid tenant for that environment:
```
{
  "security": {
    "authority": "Environment url to ChainWeb as specified by EG",
    "clientId": "Module code as defined in module manifest",
    "validEGTenantID": "Only grant access to users belonging to this EG tenant"
  }
...
}
```

Related work items: #214642